feat(generation): update request builders and models
Update generated files with build 193385 [skip ci]

+++ devicemanagement/virtual_endpoint_cloud_p_cs_item_retrieve_snapshots_request_builder.go
@@ -28,9 +28,9 @@ type VirtualEndpointCloudPCsItemRetrieveSnapshotsRequestBuilderGetQueryParameter
     // Select properties to be returned
     Select []string `uriparametername:"%24select"`
     // Skip the first n items
-    Skip *int64 `uriparametername:"%24skip"`
+    Skip *int32 `uriparametername:"%24skip"`
     // Show only the first n items
-    Top *int64 `uriparametername:"%24top"`
+    Top *int32 `uriparametername:"%24top"`
 }
 // VirtualEndpointCloudPCsItemRetrieveSnapshotsRequestBuilderGetRequestConfiguration configuration for the request such as headers, query parameters, and middleware options.
 type VirtualEndpointCloudPCsItemRetrieveSnapshotsRequestBuilderGetRequestConfiguration struct {
@@ -58,6 +58,9 @@ func NewVirtualEndpointCloudPCsItemRetrieveSnapshotsRequestBuilder(rawUrl string
 // Deprecated: This method is obsolete. Use GetAsRetrieveSnapshotsGetResponse instead.
 // returns a VirtualEndpointCloudPCsItemRetrieveSnapshotsResponseable when successful
 // returns a ODataError error when the service returns a 4XX or 5XX status code
+// [Find more info here]
+// 
+// [Find more info here]: https://learn.microsoft.com/graph/api/cloudpc-retrievesnapshots?view=graph-rest-beta
 func (m *VirtualEndpointCloudPCsItemRetrieveSnapshotsRequestBuilder) Get(ctx context.Context, requestConfiguration *VirtualEndpointCloudPCsItemRetrieveSnapshotsRequestBuilderGetRequestConfiguration)(VirtualEndpointCloudPCsItemRetrieveSnapshotsResponseable, error) {
     requestInfo, err := m.ToGetRequestInformation(ctx, requestConfiguration);
     if err != nil {
@@ -78,6 +81,9 @@ func (m *VirtualEndpointCloudPCsItemRetrieveSnapshotsRequestBuilder) Get(ctx con
 // GetAsRetrieveSnapshotsGetResponse list all cloudPcSnapshot resources for a Cloud PC.
 // returns a VirtualEndpointCloudPCsItemRetrieveSnapshotsGetResponseable when successful
 // returns a ODataError error when the service returns a 4XX or 5XX status code
+// [Find more info here]
+// 
+// [Find more info here]: https://learn.microsoft.com/graph/api/cloudpc-retrievesnapshots?view=graph-rest-beta
 func (m *VirtualEndpointCloudPCsItemRetrieveSnapshotsRequestBuilder) GetAsRetrieveSnapshotsGetResponse(ctx context.Context, requestConfiguration *VirtualEndpointCloudPCsItemRetrieveSnapshotsRequestBuilderGetRequestConfiguration)(VirtualEndpointCloudPCsItemRetrieveSnapshotsGetResponseable, error) {
     requestInfo, err := m.ToGetRequestInformation(ctx, requestConfiguration);
     if err != nil {
