+++ print/printer_shares_item_allowed_users_count_request_builder.go
@@ -44,9 +44,9 @@ func NewPrinterSharesItemAllowedUsersCountRequestBuilder(rawUrl string, requestA
 }
 // Get get the number of the resource
 // Deprecated: The printerShares navigation property is deprecated and will stop returning data on July 31, 2023. Please use the shares navigation property instead of this. as of 2023-06/Tasks_And_Plans on 2023-06-13 and will be removed 2023-07-31
-// returns a *int64 when successful
+// returns a *int32 when successful
 // returns a ODataError error when the service returns a 4XX or 5XX status code
-func (m *PrinterSharesItemAllowedUsersCountRequestBuilder) Get(ctx context.Context, requestConfiguration *PrinterSharesItemAllowedUsersCountRequestBuilderGetRequestConfiguration)(*int64, error) {
+func (m *PrinterSharesItemAllowedUsersCountRequestBuilder) Get(ctx context.Context, requestConfiguration *PrinterSharesItemAllowedUsersCountRequestBuilderGetRequestConfiguration)(*int32, error) {
     requestInfo, err := m.ToGetRequestInformation(ctx, requestConfiguration);
     if err != nil {
         return nil, err
@@ -54,14 +54,14 @@ func (m *PrinterSharesItemAllowedUsersCountRequestBuilder) Get(ctx context.Conte
     errorMapping := i2ae4187f7daee263371cb1c977df639813ab50ffa529013b7437480d1ec0158f.ErrorMappings {
         "XXX": i20a3050780ee0b0cde0a884a4f35429a20d60067e3bcda382ec5400079147459.CreateODataErrorFromDiscriminatorValue,
     }
-    res, err := m.BaseRequestBuilder.RequestAdapter.SendPrimitive(ctx, requestInfo, "int64", errorMapping)
+    res, err := m.BaseRequestBuilder.RequestAdapter.SendPrimitive(ctx, requestInfo, "int32", errorMapping)
     if err != nil {
         return nil, err
     }
     if res == nil {
         return nil, nil
     }
-    return res.(*int64), nil
+    return res.(*int32), nil
 }
 // ToGetRequestInformation get the number of the resource
 // Deprecated: The printerShares navigation property is deprecated and will stop returning data on July 31, 2023. Please use the shares navigation property instead of this. as of 2023-06/Tasks_And_Plans on 2023-06-13 and will be removed 2023-07-31
